+++ docs/changelog.rst
@@ -14,6 +14,9 @@ Changelog
   fields, ``artist_sort`` and ``albumartist_sort``, that contain sortable artist
   names like "Beatles, The". These fields are also used to sort albums and items
   when using the ``list`` command. Thanks to Paul Provost.
+* :doc:`/plugins/chroma`: The Chromaprint fingerprint and Acoustid ID are now
+  stored for all fingerprinted tracks. This version of beets *requires* at least
+  version 0.6 of `pyacoustid`_ for fingerprinting to work.
 * New :doc:`/plugins/rdm`: Randomly select albums and tracks from your library.
   Thanks to Philippe Mongeau.
 * The :doc:`/plugins/mbcollection` by Jeffrey Aylesworth was added to the core
@@ -26,6 +29,8 @@ Changelog
   deprecated ``playbin``.
 * Filenames are normalized with Unicode Normal Form D (NFD) on Mac OS X and NFC
   on all other platforms.
+
+.. _pyacoustid: https://github.com/sampsyo/pyacoustid
 
 
 1.0b13 (March 16, 2012)